Refactored resource framework.
All resources now share a common base class called BaseResource, which provides
generic methods to put and get items from the resource. This works with the old
Resource class too, which basically was a lock. Resource.request() is equal to
putting a token in the resource while Resource.release() just removes this
token again.

put() and get() accept arbitrary arguments which will be directly forwarded
into the respective event instances. This allows to specify for example custom
filters.