Do not use assert for control flow
Repair BaseResource._trigger_get() and _trigger_put() to avoid using
`assert` for control flow. Regular `if` tests are now used and
RuntimeError is raised if the assert predicate is False.

The use of `assert` for control flow caused breakage when python
optimizations were turned on. For example, this was breaking tests:

    $ PYTHONOPTIMIZE=1 py.test

This repairs issue #78.