Update testing recommendations
Update documentation in README.txt and installation.rst to recommend
running the test suite with `py.test` instead of the `python -c ...`
method. The `py.test` method is shorter and more powerful.

Plus if tests are invoked with the `py.test` method, then we no longer
need simpy.util.test(). So we also remove that function in the name of
PEP-20 (one obvious way to do it).